Simplified API for better memory usage
