+++ lib/koine/profiler/entry.rb
@@ -1,19 +1,24 @@
+# frozen_string_literal: true
+
 module Koine
   class Profiler
     class Entry
-      attr_reader :name, :elapsed_time
+      attr_reader :name
+      attr_reader :elapsed_time
+      attr_reader :memory_used
+      attr_reader :hits
 
-      def initialize(name, elapsed_time)
-        @name = name.to_s
-        @elapsed_time = elapsed_time
+      def initialize(name)
+        @name = name
+        @elapsed_time = 0
+        @memory_used = 0
+        @hits = 0
       end
 
-      def ==(other)
-        other.name == name && other.elapsed_time == elapsed_time
-      end
-
-      def <=>(other)
-        elapsed_time <=> other.elapsed_time
+      def increment(elapsed_time:, memory_used:)
+        @hits += 1
+        @elapsed_time += elapsed_time
+        @memory_used += memory_used
       end
     end
   end
